feat(node): clean up logic for overriding sigint
